coycoin_bot/requirements.txt
python-dotenv ~= 0.20.0
aiogram ~= 2.20